requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
backports.zoneinfo==0.2.1
billiard==4.2.1
cachetools==5.3.3
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-default-group==1.2.4
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==42.0.7
dj-database-url==2.2.0
Django==4.2.16
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-heroku==0.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
everapi==0.1.1
exceptiongroup==1.2.2
frozenlist==1.4.1
geoip2==4.8.0
git-filter-repo==2.38.0
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
gunicorn==23.0.0
httplib2==0.22.0
hypothesis==6.102.4
idna==3.7
iniconfig==2.0.0
ipbase==0.1
ipinfo==5.0.1
kombu==5.4.2
markdown-it-py==3.0.0
maxminddb==2.6.1
mdurl==0.1.2
multidict==6.0.5
oauthlib==3.2.2
packaging==24.0
pillow==10.4.0
pluggy==1.5.0
prompt-toolkit==3.0.48
proto-plus==1.24.0
protobuf==5.28.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.4
pyperclip==1.8.2
pytest==8.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.0.8
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.5.0
tomli==2.0.1
twilio==9.0.5
typing-extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
yarl==1.9.4
